Fix plugin loading yaml on systems where size_t is not 32-bits
git-svn-id: https://code.ros.org/svn/ros-pkg/pkg/trunk/stacks/visualization@22746 eb33c2ac-9c88-4c90-87e0-44a10359b0c3

+++ src/rviz/plugin/plugin.cpp
@@ -102,7 +102,7 @@ void Plugin::loadDescription(const std::string& description_path)
     doc["name"] >> name_;
 
     const YAML::Node& displays = doc["displays"];
-    for (size_t i = 0; i < displays.size(); ++i)
+    for (uint32_t i = 0; i < displays.size(); ++i)
     {
       const YAML::Node& n = displays[i];
       DisplayTypeInfoPtr info(new DisplayTypeInfo);
